Change return type of getFaceCount method from float to int

+++ src/core/font.h
@@ -72,7 +72,7 @@ namespace sunaba::core {
             return font->get_descent(fontSize);
         }
 
-        float getFaceCount() {
+        int getFaceCount() {
             return font->get_face_count();
         }
